Rename node_follower.rs
