+++ crates/sui-core/src/node_sync/mod.rs
@@ -1,5 +1,5 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-mod node_follower;
-pub use node_follower::*;
+mod node_state;
+pub use node_state::*;
